Move runtime struct into runtime module
